Refactor login screen and assets
- Updated AppIcons to include mail, lock, and eye logos.
- Removed mail, lock, and eye logos from AppImages.
- Added MainLayoutScreen to app routes.
- Refactored LoginScreen to use Image.asset for app logo and replaced LoginBody with LoginForm.
- Cleaned up imports in LoginScreen and other related files.
- Deleted unused LoginBody widget.
- Updated LoginButton to use constant padding and font size.
- Refactored LoginForm to use AppIcons for input field icons.
- Added mail.svg asset for email icon.
