+++ lib/core/assets/app_images.dart
@@ -8,7 +8,4 @@ class AppImages {
   static const String splashScreen = 'assets/images/splash_screen.png';
   static const String splashLogo = 'assets/images/splash_logo.png';
   static const String appLogo = 'assets/images/app_logo.png';
-  static const String mailLogo = 'assets/svgs/mail.svg';
-  static const String lockLogo = 'assets/svgs/lock.svg';
-  static const String eyeLogo = 'assets/svgs/eye.svg';
 }
