+++ lib/features/login/presentation/view/widget/login_form.dart
@@ -1,14 +1,15 @@
 import 'package:easy_localization/easy_localization.dart';
+import 'package:fitness_app/core/assets/app_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/svg.dart';
 
 import '../../../../../core/assets/app_colors.dart';
-import '../../../../../core/assets/app_images.dart';
 import '../../../../../core/utils/l10n/locale_keys.g.dart';
 import '../../../../../core/utils/shared_widgets/blured_container.dart';
 import '../../view_model/cubit/login_cubit.dart';
 import 'dont_have_account.dart';
+import 'forget_password.dart';
 
 class LoginForm extends StatelessWidget {
   const LoginForm({super.key});
@@ -43,7 +44,7 @@ class LoginForm extends StatelessWidget {
                   prefixIcon: Padding(
                     padding: const EdgeInsets.all(16),
                     child: SvgPicture.asset(
-                      AppImages.mailLogo,
+                      AppIcons.mailLogo,
                       width: 24,
                       height: 24,
                     ),
@@ -64,7 +65,7 @@ class LoginForm extends StatelessWidget {
                   prefixIcon: Padding(
                     padding: const EdgeInsets.all(16),
                     child: SvgPicture.asset(
-                      AppImages.lockLogo,
+                      AppIcons.lockLogo,
                       width: 24,
                       height: 24,
                     ),
@@ -72,7 +73,7 @@ class LoginForm extends StatelessWidget {
                   suffixIcon: Padding(
                     padding: const EdgeInsets.all(16),
                     child: SvgPicture.asset(
-                      AppImages.eyeLogo,
+                      AppIcons.eyeLogo,
                       width: 24,
                       height: 24,
                     ),
@@ -82,6 +83,8 @@ class LoginForm extends StatelessWidget {
             ],
           ),
           const SizedBox(height: 24),
+          ForgotPasswordButton(viewModel),
+          const SizedBox(height: 16),
           SizedBox(
             width: double.infinity,
             height: 50,
